requirements.txt
numpy
av
Pillow
pyTelegramBotAPI